+++ include/utils/pixel.h
@@ -8,7 +8,11 @@
 #ifndef _UTILS_PIXEL_H_
 #define _UTILS_PIXEL_H_
 
+void bgr2rgb24(unsigned char *image, unsigned int w, unsigned int h);
+void bgra2rgba32(unsigned char *image, unsigned int w, unsigned int h);
 void argb2rgba32(unsigned char *image, unsigned int w, unsigned int h);
+
+unsigned char *scaleBuffer(unsigned char *image, int width, int height, int components);
 
 #endif