+++ src/utils/tga.cpp
@@ -15,6 +15,30 @@
 #include "global.h"
 #include "utils/tga.h"
 
+typedef enum {
+    //TGA_TYPE__NO_DATA    = 0,
+    //TGA_TYPE__MAPPED     = 1,
+    TGA_TYPE__COLOR      = 2,
+    TGA_TYPE__GREYSCALE  = 3,
+    //TGA_TYPE__MAPPED_RLE = 9,
+    TGA_TYPE__COLOR_RLE  = 10
+} tga_type_t;
+
+typedef struct {
+    unsigned char comment_lenght;   //!< Number of bytes in comment
+    unsigned char colormap_type;    //!< 0 No colormap; 1 Has colormap
+    unsigned char image_type;       //!< 1 Colormapped; 2 Unmapped; 9 Colormapped RLE; 10 Unmapped RLE
+    unsigned short colormap_index;  //!< Index of first color map entry
+    unsigned short colormap_lenght; //!< Number of color map entries
+    unsigned char colormap_bbp;     //!< 16, 24, or 32 bits per pixel format
+    unsigned short origin_x;        //!< X coor of lower-left corner
+    unsigned short origin_y;        //!< Y coor of lower-left corner
+    unsigned short width;           //!< Width in pixels
+    unsigned short height;          //!< Height in pixels
+    unsigned char bpp;              //!< Number of bits in a pixel index
+    unsigned char desc_flags;       //!< Various magic bits
+} tga_t;
+
 int tgaCheck(FILE *f) {
     char buffer[10];
 
@@ -353,30 +377,5 @@ int tgaSave(FILE *f, unsigned char *image, unsigned int width, unsigned int heig
         return -2;
     }
     return 0;
-}
-
-int tgaSaveFilename(unsigned char *image, unsigned int width, unsigned int height, char type, const char *s, ...) {
-    char buffer[1024];
-    FILE *f;
-    int v;
-    va_list args;
-
-    assert(image != NULL);
-    assert(width > 0);
-    assert(height > 0);
-    assert(s != NULL);
-    assert(s[0] != '\0');
-
-    va_start(args, s);
-    vsnprintf(buffer, 1023, s, args);
-    va_end(args);
-    f = fopen(buffer, "wb");
-    if (!f) {
-        perror(buffer);
-        return -1;
-    }
-    v = tgaSave(f, image, width, height, type);
-    fclose(f);
-    return v;
 }